refactor files and edit css
