--- src/assets/css/all.scss
@@ -1,7 +1,0 @@
-@import 'base/settings';
-@import 'base/base';
-@import "components/filter";
-@import "components/profileSummary";
-@import "components/charts";
-@import "components/repos";
-@import "components/pagination";
