--- src/assets/css/base/_base.scss
@@ -1,6 +1,0 @@
-/* Reset */
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
