--- src/assets/css/components/_repos.scss
@@ -1,144 +1,0 @@
-.repos-title{
-  line-height: 4rem;
-  font-family: $font-family-base;
-  font-size: $very-large-font-size;
-  border-bottom: 4px dashed $background;
-  color: $background;
-  max-width: 14%;
-
-}
-
-.repositories{
-  padding-top: 20px;
-  padding-left: 10%;
-  background-color: $optional-background;
-  padding-bottom: 10%;
-
-}
-
-.repo-cards{
-  padding-top: 2%;
-  display: grid;
-  grid-template-columns: 30% 30% 30%;
-  grid-column-gap: 1em;
-  grid-row-gap: 1em;
-
-
-
-}
-
-.card-items{
-  display: flex;
-  background-color: #ffffff;
-  min-height: 200px;
-  box-shadow: 0 4px 8px 0 rgba(0,0,0,0.2);
-  transition: 0.3s;
-  font-family: $font-family-base;
-  flex-direction: column;
-  border-radius: 2%;
-}
-
-.card-items:hover{
-  box-shadow: 0 8px 16px 0 rgba(0,0,0,0.2);
-  cursor: pointer;
-}
-
-.repo-link{
-  text-decoration: none;
-  color: $background;
-}
-
-
-
-.repo-top{
-  min-height: 150px;
-}
-
-.spacing{
-  display: grid;
-  grid-template-columns: 10% 80% 10%;
-
-}
-
-.repo-lower-section{
-  display: flex;
-  justify-content: space-between;
-  font-size: $small-font-size;
-
-}
-
-.repo-name{
-  font-size: $large-font-size;
-  padding-bottom: 1rem;
-  padding-top: 1rem;
-
-}
-
-.repo-description{
-  font-size: $small-font-size;
-
-}
-
-.left-section{
-  position: relative;
-  bottom: 0;
-
-}
-
-.dot{
-  height: 10px;
-  width: 10px;
-  border-radius: 50%;
-  display: inline-block;
-}
-
-.repo-octicon .octicon{
-  margin-bottom: 6px;
-}
-
-@media only screen and (max-width: 800px) {
-
-  .repo-octicon .octicon{
-    margin-bottom: 2px;
-  }
-
-  .repositories{
-    margin-left: -10px;
-  }
-  .repo-cards{
-    display: grid;
-    grid-template-columns: 47% 47%;
-
-  }
-
-  .repo-name {
-    font-size: $medium-font-size;
-  }
-
-  .repo-description{
-    font-size: $very-small-font-size;
-
-  }
-
-}
-
-
-@media only screen and (max-width: 500px) {
-  .repo-octicon .octicon{
-    margin-bottom: 0;
-  }
-  .repo-name {
-    font-size: $small-font-size;
-  }
-
-  .repo-description{
-    font-size: $very-small-font-size;
-
-  }
-  .repo-cards{
-    display: grid;
-    grid-template-columns: 95%;
-
-  }
-
-}
