+++ src/helpers/barChart.ts
