+++ src/index.css
@@ -1,0 +1,6 @@
+/* Reset */
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+}
